requirements.txt
asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
model-bakery==1.17.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.5.2
pytest-testdox==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.6.0
